Rename modules
Rename some modules for consistency

* checkout_externals - removed .py extension
* externals_status - consistency with exernals_description
* global_constants - clarify intent
* test_unit_externals_status - consistency with module being tested.

Testing:
  python2 - unit tests - pass
  python2 - manual testing cesm - ok
  python2 - manual testing clm - ok
