+++ manic/externals_status.py
@@ -8,7 +8,7 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 from __future__ import print_function
 
-from .globals import EMPTY_STR
+from .global_constants import EMPTY_STR
 from .utils import printlog
