+++ manic/global_constants.py
